IDK (I Don't Know)
This is a custom web application made at and for Apple Retail in 2008.
It was made to provide a quick reference for Specialists to common
questions on the sales floor.  It was made to run on an iPhone with a
native look and at speeds.  Although this iteration was not deployed on
a large scale, it’s base code, UI/UX design was the foundation for the
Smart Signs that everyone uses to answer common questions about Apple
goods and services.